requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
tensorflow==2.15.0
Pillow==10.4.0
numpy>=1.26.4
